Handle missing log/xml files correctly
The testrunner script failed and exited immediately
when trying to rename a non-existent file. This
patch makes sure the file is checked to be existing
before trying to move it.

Closes-Bug: #1804165

+++ tntestr/cmd/tntestr.py
@@ -205,8 +205,15 @@ class TungstenTestRunner(object):
             if not (os.path.exists(xml_path) or os.path.exists(log_path)):
                 break
 
-        os.rename(suite['xml_path'], xml_path)
-        os.rename(suite['log_path'], log_path)
+        if os.path.exists(suite['xml_path']):
+            os.rename(suite['xml_path'], xml_path)
+        else:
+            logging.warning('{} does not exist!'.format(suite['xml_path']))
+
+        if os.path.exists(suite['log_path']):
+            os.rename(suite['log_path'], log_path)
+        else:
+            logging.warning('{} does not exist!'.format(suite['log_path']))
 
         result_text = "SUCCESS" if result == TestResult.SUCCESS else "FAILURE"
         if result > TestResult.SUCCESS:
